login_required added

+++ storeregister/views.py
@@ -2,6 +2,7 @@ from django.shortcuts import render, get_object_or_404, redirect
 from django.contrib.auth.decorators import login_required
 from .models import Store
 from .forms import StoreForm
+from accounts.models import User
 
 # Create your views here.
 
@@ -16,44 +17,53 @@ def detail(request, store_id):
     return render(request, 'storeregister/store_detail.html', {'store':store})
 
 ##스토어 등록 (C)
-#@login_required
+@login_required
 def register(request):
     if request.method == 'POST':
         form = StoreForm(request.POST, request.FILES)
         if form.is_valid():
-            post = form.save(commit=False)
-            post.save()
+            store = form.save(commit=False)
+            store.user = request.user  ##username 자동 설정
+            store.save()
             return redirect('list')
-
     else : 
         form=StoreForm()
         return render(request, 'storeregister/store_register.html', {'form':form})
 
 
 ##스토어 수정(U)
-#@login_required
+@login_required
 def storeupdate(request, store_id):
     store = get_object_or_404(Store, pk=store_id)
     if request.method=='POST':
         form = StoreForm(request.POST, instance=store)
         if form.is_valid():
             store = form.save(commit=False)
+            store.user = request.user  ##username 자동 설정
             store.save()
             return redirect('detail', store_id=store.pk)    
     else:
-        form = StoreForm(instance=store)
-        return render(request, 'storeregister/store_edit.html', {'form':form})
+        if store.user == User.objects.get(username=request.user.get_username()) :  ##자신의 글일때만 수정 가능
+            form = StoreForm(instance=store)
+            return render(request, 'storeregister/store_edit.html', {'form':form})
+
+        else :
+            return redirect('detail', store_id=store.pk)    ##자신의 글 아니면 해당글 detail로 redirect
 
 
 ##스토어 삭제(D)
-#@login_required
+@login_required
 def delete(request, store_id):
     store = get_object_or_404(Store, pk=store_id)
-    store.delete()
-    return redirect('list')
+    if store.user == User.objects.get(username=request.user.get_username()) : ##자신의 글일때만 삭제 가능
+        store.delete()
+        return redirect('list')
+    else:
+        return redirect('detail', store_id=store.pk)   ##자신의 글 아니면 해당글 detail로 redirect
 
 
 
+#필터링
 def result(request) : 
     store_object = Store.objects
     query = request.GET.get('query','')
